conversations & messages middleware fertig

+++ controller/messagesController.js
@@ -6,6 +6,7 @@ import jwt from "jsonwebtoken";
 // I M P O R T:  F U N C T I O N S
 import MessageModel from "../models/messageModel.js";
 import ConversationModel from "../models/conversationModel.js";
+import UserModel from "../models/userModel.js";
 
 // I M P O R T  &  D E C L A R E   B C R Y P T   K E Y
 const JWT_KEY = process.env.SECRET_JWT_KEY || "DefaultValue";
@@ -36,9 +37,7 @@ export async function messagesGetAll(req, res, next) {
 export async function addMessage(req, res, next) {
   try {
     // TAKE USERID
-    const token = req.cookies.loginCookie;
-    const tokenDecoded = jwt.verify(token, JWT_KEY);
-    const userId = tokenDecoded.userId;
+    const userId = req.body.userId;
 
     // TAKE CONVERSATIONID
     const conversationId = req.body.conversationId;
@@ -67,23 +66,29 @@ export async function addMessage(req, res, next) {
 export async function updateMessage(req, res, next) {
   try {
     // const conversationId=req.body.conversationId;
-    const updatedMessage = await MessageModel.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      {
-        new: true,
-      }
-    );
-    // const pushInConversation = await ConversationModel.findByIdAndUpdate(
-    //   {conversationId },
-    //   { $push: { message: updatedMessage } }
-    // );
-
-    res.status(201).json({
-      message: "message edited successfully",
-      status: true,
-      data: updatedMessage,
-    });
+    const userId = req.body.userId;
+    const sender = await MessageModel.findById(req.params.id);
+    // nur der ersteller kann seine Nachricht lösen
+    if (sender.from.toString() === userId) {
+      const updatedMessage = await MessageModel.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        {
+          new: true,
+        }
+      );
+      res.status(201).json({
+        message: "message edited successfully",
+        status: true,
+        data: updatedMessage,
+      });
+    } else {
+      res.status(403).send({
+        message: "you are not allowed to edit or delete this message",
+        status: false,
+        data: "",
+      });
+    }
   } catch (error) {
     next(error);
   }
@@ -93,15 +98,37 @@ export async function updateMessage(req, res, next) {
 export async function deleteMessage(req, res, next) {
   try {
     const conversationId = req.body.conversationId;
-    await MessageModel.findByIdAndDelete(req.params.id);
-    await ConversationModel.findById(conversationId, {
-      $pull: { message: req.params.id },
-    });
-    res.status(201).send({
-      message: "message deleted successfully ",
-      status: true,
-      data: "",
-    });
+
+    const userId = req.body.userId;
+    const sender = await MessageModel.findById(req.params.id);
+    // nur der ersteller kann sein Nachricht lösen
+    if (sender.from.toString() === userId) {
+      await MessageModel.findByIdAndDelete(req.params.id);
+      await ConversationModel.findById(conversationId).updateOne({
+        $pull: {
+          message: req.params.id,
+        },
+      });
+      const conversation = await ConversationModel.findById(conversationId);
+      if (conversation.message.length === 0) {
+        await ConversationModel.findByIdAndDelete(conversationId);
+        await UserModel.find({ conversations: conversationId }).updateMany({
+          $pull: { conversations: conversationId },
+        });
+      }
+      console.log("message length" + conversation.message.length);
+      res.status(201).send({
+        message: "message deleted successfully ",
+        status: true,
+        data: "",
+      });
+    } else {
+      res.status(403).send({
+        message: "you are not allowed to edit or delete this message",
+        status: false,
+        data: "",
+      });
+    }
   } catch (error) {
     next(error);
   }
